+++ packages/isbt/.eslintrc.js
@@ -1,0 +1,1 @@
+module.exports = require('./eslint.config.base')
