+++ packages/isbt/package.json
@@ -9,7 +9,7 @@
     "isbt": "dist/cjs/index.js"
   },
   "files": [
-    ".eslintrc.base.js",
+    "eslint.config.base.js",
     "babel.config.js",
     "tsconfig.base.json",
     "dist/cjs",
@@ -22,6 +22,7 @@
     "@lerna/cli": "^4.0.0",
     "@lerna/exec": "^4.0.0",
     "@types/jest": "^27.0.0",
+    "@types/minimist": "^1.0.0",
     "@typescript-eslint/eslint-plugin": "^4.0.0",
     "@typescript-eslint/parser": "^4.0.0",
     "eslint": "^7.0.0",
@@ -31,6 +32,7 @@
     "husky": "^4.0.0",
     "jest": "^27.0.0",
     "lint-staged": "^10.0.0",
+    "minimist": "^1.0.0",
     "prettier": "^2.0.0",
     "tslib": "^2.0.0",
     "typescript": "^4.0.0"
